refactor(services): extract dev tool implementations to dev_tools.py
Dashboard YAML generator (_build_* / _format_* / channel helpers /
create_generate_dashboard_handler) and channel identity converter
(_build_channel_lookup / _plan_stat_renames_* / _migrate_statistics /
create_convert_channel_identity_handler) move to a new dev_tools.py.

services.py becomes a thin registration layer (235 lines) importing
handler factories from dev_tools. Retains service constants, schemas,
async_request_modem_refresh, _find_loaded_entries / _resolve_target_entries,
request_refresh / request_health_check handler factories, and
async_register/unregister_services.

Test imports and mock patch targets (services.dr, services._resolve_config_entry_for_device)
updated to dev_tools. No logic changes. 100% coverage on both files.

Spec updates: HA_ADAPTER_SPEC.md Module Inventory corrected,
BLUEPRINT_DISTRIBUTION_SPEC.md Out of Scope note updated,
CODE_REVIEW.md source→test mapping extended.
